% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD_evaluation.R
\name{AD_evaluation}
\alias{AD_evaluation}
\title{Average Difference (AD) for Evaluation the different filters}
\usage{
AD_evaluation(image, filtered_image, filter_name = NULL)
}
\arguments{
\item{image}{A numeric matrix, raster::RasterLayer, or terra::SpatRaster representing the image.}

\item{filtered_image}{filtered image}

\item{filter_name}{The name of the filter, which is used}
}
\value{
Average Difference
}
\description{
Calculate the Average Difference to measure, which filter has the best outcome.
}
\details{
Source: Comparison of Various Speckle Noise Reduction Filters on Synthetic Aperture Radar Image.Ardhi Wicaksono Santoso, Luhur Bayuaji, Lim Tien Sze, Habibah Lateh, Jasni Mohamad Zain
}
\examples{
img <- load_example()
filters <- all_filters(img, window_size = 3, ENL = 3.5, plot_result = FALSE)
ad_metrics <- AD_evaluation(img, filters)
print(ad_metrics)
}
