% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_filters.R
\name{all_filters}
\alias{all_filters}
\title{All Filters (mean, median, kuan and lee) for Speckle Noise}
\usage{
all_filters(image, window_size = 3, ENL = NULL, plot_result = TRUE)
}
\arguments{
\item{image}{A numeric matrix, raster::RasterLayer, or terra::SpatRaster representing the image.}

\item{window_size}{An integer specifying the size of the window. Default is 3.}

\item{ENL}{Effective Number of Looks.}

\item{plot_result}{If TRUE, plots Original + filtered images.}
}
\value{
A named list with filtered images: $lee, $kuan, $mean, $median
}
\description{
Applies all filters to a grayscale image/raster to reduce speckle noise.
}
