% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ENL.R
\name{estimate_ENL}
\alias{estimate_ENL}
\title{Estimate ENL interactively from a satellite raster}
\usage{
estimate_ENL(raster)
}
\arguments{
\item{raster}{A numeric matrix, RasterLayer, or terra::SpatRaster.}
}
\value{
A numeric value representing the estimated ENL.
}
\description{
Allows the user to select a homogeneous, flat area in a satellite image
to estimate the Equivalent Number of Looks (ENL) for speckle filtering.
Click multiple points to define a polygon and press ESC when done.
}
\examples{
\dontrun{
img <- load_example()
enl <- estimate_ENL(img)  # interactive: select region with mouse
}
}
