% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI_evaluation.R
\name{SI_evaluation}
\alias{SI_evaluation}
\title{Speckle Index (S)) for Evaluation the different filters}
\usage{
SI_evaluation(image, filtered_image, radius = 1, filter_name = NULL)
}
\arguments{
\item{image}{A numeric matrix, raster::RasterLayer, or terra::SpatRaster representing the image.}

\item{filtered_image}{filtered image}

\item{filter_name}{The name of the filter, which is used}
}
\value{
Speckle Index
}
\description{
Calculate the Speckle Index to measure, which filter has the best outcome.
}
\details{
Source: Comparison of Various Speckle Noise Reduction Filters on Synthetic Aperture Radar Image.Ardhi Wicaksono Santoso, Luhur Bayuaji, Lim Tien Sze, Habibah Lateh, Jasni Mohamad Zain
#'
}
\examples{
img <- load_example()
filters <- all_filters(img, window_size = 3, ENL = 3.5, plot_result = FALSE)
si_metrics <- SI_evaluation(img, filters)
print(si_metrics)
}
