% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNR_evaluation.R
\name{SNR_evaluation}
\alias{SNR_evaluation}
\title{Signal to Noise Ratio (SNR) for Evaluation}
\usage{
SNR_evaluation(image, filtered_image, filter_name = NULL)
}
\arguments{
\item{image}{A numeric matrix, raster::RasterLayer, or terra::SpatRaster representing the image.}

\item{filtered_image}{filtered image}

\item{filter_name}{The name of the filter, which is used}
}
\value{
Signal to Noise Ratio
}
\description{
Calculate the Signal to Noise Ratio (SNR) to measure, which filter has the best outcome.
}
\details{
Source: Comparison of Various Speckle Noise Reduction Filters on Synthetic Aperture Radar Image.Ardhi Wicaksono Santoso, Luhur Bayuaji, Lim Tien Sze, Habibah Lateh, Jasni Mohamad Zain
#'
}
\examples{
img <- load_example()
filters <- all_filters(img, window_size = 3, ENL = 3.5, plot_result = FALSE)
snr_metrics <- SNR_evaluation(img, filters)
print(snr_metrics)
}
