% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_filters.R
\name{all_filters}
\alias{all_filters}
\title{All Filters (mean, median, kuan and lee) for Speckle Noise}
\usage{
all_filters(image, window_size = 3, ENL = NULL, plot_result = TRUE)
}
\arguments{
\item{image}{A numeric matrix, raster::RasterLayer, or terra::SpatRaster representing the image.}

\item{window_size}{An integer specifying the size of the local window for median/mean filters. Default is 3.}

\item{ENL}{Effective Number of Looks. If NULL, it is estimated per filter or you can use the function estimae_enl()}

\item{plot_result}{Logical. If TRUE, plots Original + filtered images.}
}
\value{
A named list with filtered images: $lee, $kuan, $mean, $median
}
\description{
Applies all filters to a grayscale image or raster to reduce speckle noise.
}
\examples{
\dontrun{
img <- load_example()
filters <- all_filters(img, window_size = 3, ENL = 3.5, plot_result = TRUE)
}
}
