% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lee_filter.R
\name{lee_filter}
\alias{lee_filter}
\title{Lee Filter for Speckle Noise}
\usage{
lee_filter(image, window_size = 3, ENL = NULL)
}
\arguments{
\item{image}{A numeric matrix, raster::RasterLayer, or terra::SpatRaster representing the image.}

\item{window_size}{An odd integer specifying the size of the local window. Default is 3.}

\item{ENL}{Equivalent number of looks.}
}
\value{
The filtered image. Returns a matrix if input was a matrix, RasterLayer if input was raster::RasterLayer, or SpatRaster if input was terra::SpatRaster.
}
\description{
Applies a Lee filter to a grayscale image or raster to reduce speckle noise.
}
\details{
Source: Active Remote Sensing - Tobi Ullmann
}
\examples{
\dontrun{
img <- load_example()
lee <- lee_filter(img, window_size = 3, ENL = 3.5)
}
}
