% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example.R
\name{load_example}
\alias{load_example}
\title{Load Example Sentinel Image}
\usage{
load_example()
}
\value{
A SpatRaster object
}
\description{
Loads the example Sentinel-1 TIF included in the package.
}
\examples{
img <- load_example()
plot(img)
}
