% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_filter.R
\name{median_filter}
\alias{median_filter}
\title{Median Filter for Speckle Noise}
\usage{
median_filter(image, window_size = 3)
}
\arguments{
\item{image}{A numeric matrix, raster::RasterLayer, or terra::SpatRaster representing the image.}

\item{window_size}{An integer specifying the size of the median filter window. Default is 3.}
}
\value{
The filtered image. Returns a matrix if input was a matrix, RasterLayer if input was raster::RasterLayer, or SpatRaster if input was terra::SpatRaster.
}
\description{
Applies a median filter to a grayscale image or raster to reduce speckle noise.
}
\examples{
\dontrun{
img <- load_example()
median_f <- median_filter(img, window_size = 3)
}
}
