% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_evaluation.R
\name{MSE_evaluation}
\alias{MSE_evaluation}
\title{Mean Square Error for Evaluation the different Filter}
\usage{
MSE_evaluation(image, filtered_image, filter_name = NULL)
}
\arguments{
\item{image}{A numeric matrix, raster::RasterLayer, or terra::SpatRaster representing the image.}

\item{filtered_image}{filtered image}

\item{filter_name}{The name of the filter, which is used}
}
\value{
Mean Square Error
}
\description{
Calculate the Mean Square Error to measure, which filter has the best outcome.
}
\details{
Source: Comparison of Various Speckle Noise Reduction Filters on Synthetic Aperture Radar Image.Ardhi Wicaksono Santoso, Luhur Bayuaji, Lim Tien Sze, Habibah Lateh, Jasni Mohamad Zain
#'
}
\examples{
img <- load_example()
filters <- all_filters(img, window_size = 3, ENL = 3.5, plot_result = FALSE)
mse_metrics <- MSE_evaluation(img, filters)
print(mse_metrics)
}
